RHI перестал говорить на языке OpenGL, профилировщик обрёл читателя
Три архитектурные правки, у каждой своя болезнь.

ХЕНДЛЫ RHI были сырым unsigned int — именем объекта OpenGL, торчащим
наружу через весь пост-процесс, тени и отражения. Это не стилистика: у
Vulkan и Metal хендл — 64-битный указатель, в 32 бита он не влезает, и
второй бэкенд упёрся бы не в архитектуру, а в арифметику. Хуже была
вторая половина: число без типа путается с чем угодно, и хендл текстуры,
хендл запроса и номер юнита компилятор различить не мог. Теперь это
TextureHandle и QueryHandle без неявного приведения к числу; сырое число
осталось в одной отдельно названной двери — NativeHandle() для ImGui,
потому что передача текстуры в стороннее API бэкенд-специфична по своей
природе, и притворяться иначе было бы хуже протечки.

Заодно записаны СОГЛАШЕНИЯ, о которых интерфейс молчал, а бэкенд решал по
привычкам GL: начало отсчёта прямоугольников, порядок строк ReadPixels,
начало координат текстуры, диапазон глубины NDC. Молчание тут опаснее
протечки типа — оно не ломает сборку, а даёт перевёрнутую картинку на
втором бэкенде, и ищется это неделями.

НАБОР СООТВЕТСТВИЯ (sage/rhi/Conformance.h) — один исполняемый контракт,
который обязан пройти каждый бэкенд. Пока реализация одна, «интерфейс» и
«то, что делает OpenGL» неразличимы. Два уровня: структурный проходят все,
включая Null; пиксельный — только те, кто растеризует. Проверено, что он
не пустой: перевёрнутое начало отсчёта ножниц в GL-бэкенде роняет его,
хотя ни один из 97 эталонных кадров этого не замечает. Это сужает
гипотезу, а не снимает её: Vulkan по-прежнему потребует переписать модель
интерфейса, и в README это сказано прямо.

ПРОФИЛИРОВЩИК был готов целиком — метки GPU, отложенное чтение, честный
отказ на драйвере без таймеров, — и не вызывался НИ РАЗУ: ни одного
SetEnabled(true), ни одного читателя. Теперь размечен весь кадр (каждая
система планировщика по имени, отсечение, тени, геометрия, скиннинг,
отражения, интерфейс и каждый проход пост-обработки), появилась панель
Window -> Профилировщик и переменные SAGE_PROFILE / SAGE_PROFILE_LOG для
машин без окна. Первый же прогон боевой игры показал, что пост-обработка
стоит больше геометрии, а половину внутри неё съедает SSAO.

COMPONENTS.H был монолитом на 583 строки, где соседствовали интерфейс, IK,
зонды отражений, физика, частицы и анимация, и рос при каждой фиче.
Компоненты разложены по своим системам (десять самодостаточных
заголовков), Components.h остался зонтиком — три десятка включений
работают как раньше.

229 юнит-тестов, 99 рендер-тестов, 7/7 smoke, граница RHI чиста.

+++ editor/src/EditorHost.h
@@ -1,4 +1,5 @@
 #pragma once
+#include <cstdint>
 #include <filesystem>
 #include <string>
 #include <vector>
@@ -109,13 +110,16 @@ public:
     virtual Camera& EditorCamera() = 0;
     virtual const glm::mat4& ViewMatrix() const = 0;
     virtual const glm::mat4& ProjMatrix() const = 0;
-    virtual unsigned int SceneTexture() const = 0;     // цвет FBO сцены (для ImGui::Image)
+    // Сырое число бэкенда для ImGui::Image — единственная легальная дверь
+    // наружу (см. rhi::Texture2D::NativeHandle). Панель показывает кадр сцены
+    // сторонним API, и это бэкенд-специфично по своей природе.
+    virtual uint64_t SceneTexture() const = 0;
     virtual void SetViewportSize(int w, int h) = 0;    // панель сообщает размер под FBO
     // u,v в [0..1] — выбор сущности лучом. additive (Ctrl) — добавить/убрать из набора.
     virtual void PickAtViewport(float u, float v, bool additive = false) = 0;
 
     // --- панель Game (игровое окно: рендер от Primary-камеры сцены) ---
-    virtual unsigned int GameTexture() const = 0;
+    virtual uint64_t GameTexture() const = 0;
     virtual void SetGameViewportSize(int w, int h) = 0;
     virtual bool HasPrimaryCamera() = 0;
